refactor(diyui): 拆分 panel provider 组件到独立文件
按 Panel 目录习惯分类:
  - _base.py: UIComponent, _PanelContainerMixin 基础类
  - layout/: PanelColumn, PanelRow, PanelCard
  - pane/: PanelMarkdown
  - widgets/: PanelButton, PanelTextInput, PanelRadioButtonGroup

panel.py 保留为向后兼容入口（re-export *）。
36 个测试全部通过。
