+++ packages/diyui.py/src/diyui/providers/panel/pane/__init__.py
@@ -1,0 +1,3 @@
+from diyui.providers.panel.pane.markdown import PanelMarkdown
+
+__all__ = ["PanelMarkdown"]
