+++ packages/diyui.py/src/diyui/providers/panel/widgets/button.py
@@ -1,0 +1,86 @@
+"""PanelButton — pn.widgets.Button 的 diyui 包装。"""
+
+from __future__ import annotations
+
+from typing import Any
+
+import panel as pn
+
+from diyui.providers.panel._base import UIComponent
+
+
+class PanelButton(UIComponent, pn.widgets.Button):
+    """Panel Button 包装，同时是 pn.widgets.Button 实例。"""
+
+    def __init__(
+        self,
+        *,
+        label: str = "",
+        name: str = "",
+        value: bool = False,
+        align: Any = "start",
+        aspect_ratio: Any | None = None,
+        css_classes: list[Any] | None = None,
+        design: Any = None,
+        height: int | None = None,
+        min_width: int | None = None,
+        min_height: int | None = None,
+        max_width: int | None = None,
+        max_height: int | None = None,
+        margin: Any | None = (5, 10),
+        styles: dict[str, Any] | None = None,
+        stylesheets: list[Any] | None = None,
+        tags: list[Any] | None = None,
+        width: int | None = None,
+        width_policy: Any = "auto",
+        height_policy: Any = "auto",
+        sizing_mode: Any = None,
+        visible: bool = True,
+        loading: bool = False,
+        disabled: bool = False,
+        description: Any | None = None,
+        description_delay: int = 500,
+        icon: str | None = None,
+        icon_size: str = "1em",
+        color: Any = "default",
+        variant: Any = "solid",
+        button_type: Any | None = None,
+        button_style: Any | None = None,
+    ) -> None:
+        # label 取代 name：label 优先，name 仅在 label 未设置时作为兼容回退
+        _label = label or name
+        # color/variant 取代 button_type/button_style：新参数优先
+        _color = color if color != "default" else (button_type or "default")
+        _variant = variant if variant != "solid" else (button_style or "solid")
+        UIComponent.__init__(self)
+        pn.widgets.Button.__init__(
+            self,
+            label=_label,
+            value=value,
+            align=align,
+            aspect_ratio=aspect_ratio,
+            css_classes=css_classes or [],
+            design=design,
+            height=height,
+            min_width=min_width,
+            min_height=min_height,
+            max_width=max_width,
+            max_height=max_height,
+            margin=margin,
+            styles=styles or {},
+            stylesheets=stylesheets or [],
+            tags=tags or [],
+            width=width,
+            width_policy=width_policy,
+            height_policy=height_policy,
+            sizing_mode=sizing_mode,
+            visible=visible,
+            loading=loading,
+            disabled=disabled,
+            description=description,
+            description_delay=description_delay,
+            icon=icon,
+            icon_size=icon_size,
+            color=_color,
+            variant=_variant,
+        )
